Вынес MainWindow в отдельный файл. Создал окно для ведущего.
