+++ helpers/signals.py
@@ -4,3 +4,8 @@ from dataclasses import dataclass
 @dataclass
 class SignalArgs:
     sender: object
+
+
+@dataclass
+class OnResizeSignalArgs(SignalArgs):
+    pass
